python/requirements.txt
fastf1
pandas
numpy
psycopg[binary]
python-dotenv